Ejercicios dart

+++ trim1/programacion/dart_adso/Ejercicios_dart/Ejercicio_1.dart
@@ -4,7 +4,7 @@ otra para su género y otra para su precio en dólares.*/
 void main(List<String> args) {
   String nombre = "GTA 6";
   String genero = "Rol";
-  int precio = 500000;
+  double precio = 500000;
 
 print("El nombre del video juego es $nombre, su genero es $genero y cuesta $precio dolares" );
